# backend/openapi.yaml
openapi: 3.0.1
info:
  title: LHIND - The Gremlins API
  description: API documentation for "The Gremlins".
  contact:
    name: The Gremlins
    url: https://github.com/omega0verride
  version: "1.0"
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
  /api/user/users/{username}/profileImage:
    post:
      tags:
      - app-user-controller
      operationId: uploadImage
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - app-user-controller
      operationId: deleteImage
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/token:
    post:
      tags:
      - auth-token-controller
      operationId: authenticate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiAuthIdentityProviderDTO'
        required: true
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DetailedTokenDetailsDTO'
  /api/token/refresh:
    post:
      tags:
      - auth-token-controller
      operationId: refreshToken
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DetailedTokenDetailsDTO'
  /api/public/users/register:
    post:
      tags:
      - app-user-controller
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppUserDTO'
        required: true
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/admin/users:
    get:
      tags:
      - app-user-controller
      operationId: getUsers
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetAppUserDTO'
    post:
      tags:
      - app-user-controller
      operationId: registerUserFromAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppUserDTO'
        required: true
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/user/users/{username}:
    get:
      tags:
      - app-user-controller
      operationId: getUserByUsername
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GetAppUserDTO'
    delete:
      tags:
      - app-user-controller
      operationId: deleteUserByUsername
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    patch:
      tags:
      - app-user-controller
      operationId: updateUserByUsername
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppUserDTO'
        required: true
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GetAppUserDTO'
  /api/public/helloworld:
    get:
      tags:
      - hello-world-controller
      operationId: helloWorld
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/admin/users/{ids}:
    delete:
      tags:
      - app-user-controller
      operationId: deleteUsers
      parameters:
      - name: ids
        in: path
        required: true
        schema:
          type: array
          items:
            type: integer
            format: int64
      responses:
        "415":
          description: Unsupported Media Type
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse__'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
    ErrorResponse__:
      type: object
      properties:
        message:
          type: string
        detailed_message:
          type: string
        details:
          type: object
          additionalProperties:
            type: object
        details_list:
          type: array
          items:
            type: string
        success:
          type: boolean
    MultiAuthIdentityProviderDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        token:
          type: string
        authType:
          type: string
          enum:
          - BASIC
          - GOOGLE
          - FACEBOOK
          - INSTAGRAM
    DetailedTokenDetailsDTO:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        username:
          type: string
        expires_at:
          type: string
          format: date-time
        authorities:
          type: array
          items:
            type: string
    CreateAppUserDTO:
      required:
      - email
      - name
      - password
      - surname
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        password:
          type: string
        phoneNumber:
          type: string
        registerAsFoodProvider:
          type: boolean
        locale:
          type: object
          properties:
            language:
              type: string
            script:
              type: string
            country:
              type: string
            variant:
              type: string
            unicodeLocaleAttributes:
              uniqueItems: true
              type: array
              items:
                type: string
            unicodeLocaleKeys:
              uniqueItems: true
              type: array
              items:
                type: string
            displayLanguage:
              type: string
            displayScript:
              type: string
            displayCountry:
              type: string
            displayVariant:
              type: string
            displayName:
              type: string
            extensionKeys:
              uniqueItems: true
              type: array
              items:
                type: string
            iso3Language:
              type: string
            iso3Country:
              type: string
    UpdateAppUserDTO:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        phoneNumber:
          type: string
    GetAppUserDTO:
      type: object
      properties:
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        username:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        enabled:
          type: boolean
        profileImagePath:
          type: string